requirements.txt
pillow==10.2.0
beautifulsoup4==4.12.3
bcrypt==4.3.0
fpdf==1.7.2